Extract chart data formater
