+++ app/javascript/data/formater.js
@@ -1,0 +1,9 @@
+import { accessInfo } from './formaters/accessInfoFormater'
+import { mostServiceTracked } from './formaters/mostServiceTrackedFormater'
+import { feedbackTrend } from './formaters/feedbackTrendFormater'
+
+export default {
+  accessInfo,
+  mostServiceTracked,
+  feedbackTrend
+}
