+++ app/javascript/data/formaters/feedbackTrendFormater.js
@@ -1,0 +1,6 @@
+export const feedbackTrend = ({ ratingLabels, dataset }) => {
+  return {
+    labels: ratingLabels,
+    datasets: _.map(dataset, (el) => el)
+  };
+}
