+++ app/views/welcomes/_tabs.haml
@@ -26,7 +26,7 @@
           .col-xl-6.col-lg-5
             = render( ChartComponent.new(name: t("welcomes.information_access_by"), filterable: true) ) do |c|
               - c.with(:addon) do
-                %select.access-period
+                %select.access-period.period-filter
                   %option{ value: "month" }= t("period.monthly")
                   %option{ value: "quarter" }= t("period.quarterly")
                   %option{ value: "year" }= t("period.yearly")
@@ -40,7 +40,7 @@
             - name = t("welcomes.most_service_tracked_by_periodic")
             = render( ChartComponent.new(name: name, filterable: true) ) do |c|
               - c.with(:addon) do
-                %select.track-period
+                %select.track-period.period-filter
                   %option{ value: "month" }= t("period.monthly")
                   %option{ value: "quarter" }= t("period.quarterly")
                   %option{ value: "year" }= t("period.yearly")
@@ -69,7 +69,7 @@
           .col-xl-6.col-lg-5
             = render( ChartComponent.new(name: t("welcomes.owso_feedback_trend"), filterable: true) ) do |c|
               - c.with(:addon) do
-                %select.feedback-trend
+                %select.feedback-trend.period-filter
                   %option{ value: "month" }= t("period.monthly")
                   %option{ value: "quarter" }= t("period.quarterly")
                   %option{ value: "year" }= t("period.yearly")
